Detect stale streams by snapshot interval
If the interval between snapshots is too long, it's likely that the
tracking is not happening (perhaps the machine is asleep). If that's
the case, then the snapshot stream should be considered stale,
and Windows shouldn't not be considered to have been active in the
interval.

This commit does that. It includes a hardcoded interval of 1m. This
could easily be parameterised by creating a new option for show and
using time.ParseDuration. I haven't gotten that far :D